Rename FirstWallBuilder -> InComponentVesselBuilder
This intends to remove the ambiguity around the builder's purpose.
As the FirstWallBuilder's components are not necessarily real, physical,
components, they're more used for scaffolding in order to build the real
components in the reactor
