+++ tests/bluemira/builders/EUDEMO/ivc/test_ivc.py
@@ -30,17 +30,17 @@ import pytest
 
 from bluemira.base.error import BuilderError
 from bluemira.base.file import get_bluemira_path
-from bluemira.builders.EUDEMO.first_wall import FirstWallBuilder
-from bluemira.builders.EUDEMO.first_wall.wall import WallBuilder
+from bluemira.builders.EUDEMO.ivc import InVesselComponentBuilder
+from bluemira.builders.EUDEMO.ivc.wall import WallBuilder
 from bluemira.equilibria import Equilibrium
 from bluemira.equilibria.find import find_OX_points
 
 DATA = get_bluemira_path("bluemira/equilibria/test_data", subfolder="tests")
 OPTIMISER_MODULE_REF = "bluemira.geometry.optimisation"
-WALL_MODULE_REF = "bluemira.builders.EUDEMO.first_wall.wall"
+WALL_MODULE_REF = "bluemira.builders.EUDEMO.ivc.wall"
 
 
-class TestFirstWallBuilder:
+class TestInVesselComponentBuilder:
 
     _default_variables_map = {
         "x1": {  # ib radius
@@ -52,7 +52,7 @@ class TestFirstWallBuilder:
     }
 
     _default_config = {
-        "name": "First Wall",
+        "name": "In Vessel Components",
         "param_class": f"{WALL_MODULE_REF}::WallPrincetonD",
         "problem_class": f"{OPTIMISER_MODULE_REF}::MinimiseLength",
         "runmode": "mock",
@@ -60,7 +60,7 @@ class TestFirstWallBuilder:
     }
 
     _params = {
-        "Name": "First Wall Example",
+        "Name": "IVC Example",
         "plasma_type": "SN",
         # Wall shape opts
         "R_0": (9.0, "Input"),
@@ -81,7 +81,7 @@ class TestFirstWallBuilder:
         _, cls.x_points = find_OX_points(cls.eq.x, cls.eq.z, cls.eq.psi())
 
     def test_wall_boundary_is_cut_below_x_point_in_z_axis(self):
-        builder = FirstWallBuilder(
+        builder = InVesselComponentBuilder(
             self._params, build_config=self._default_config, equilibrium=self.eq
         )
 
@@ -95,27 +95,29 @@ class TestFirstWallBuilder:
         )
 
     def test_contains_one_divertor_component_given_SN_plasma(self):
-        builder = FirstWallBuilder(
+        builder = InVesselComponentBuilder(
             self._params, build_config=self._default_config, equilibrium=self.eq
         )
 
         wall = builder()
 
-        divertors = wall.get_component(FirstWallBuilder.COMPONENT_DIVERTOR, first=False)
+        divertors = wall.get_component(
+            InVesselComponentBuilder.COMPONENT_DIVERTOR, first=False
+        )
         assert len(divertors) == 1
 
     def test_contains_one_wall_component(self):
-        builder = FirstWallBuilder(
+        builder = InVesselComponentBuilder(
             self._params, build_config=self._default_config, equilibrium=self.eq
         )
 
         wall = builder()
 
-        walls = wall.get_component(FirstWallBuilder.COMPONENT_WALL, first=False)
+        walls = wall.get_component(InVesselComponentBuilder.COMPONENT_WALL, first=False)
         assert len(walls) == 1
 
     def test_component_tree_structure(self):
-        builder = FirstWallBuilder(
+        builder = InVesselComponentBuilder(
             self._params, build_config=self._default_config, equilibrium=self.eq
         )
 
@@ -125,16 +127,16 @@ class TestFirstWallBuilder:
         xz = first_wall.get_component("xz")
         assert xz is not None
         assert xz.depth == 1
-        divertor_xz = xz.get_component(FirstWallBuilder.COMPONENT_DIVERTOR)
+        divertor_xz = xz.get_component(InVesselComponentBuilder.COMPONENT_DIVERTOR)
         assert divertor_xz.depth == 2
-        wall_xz = xz.get_component(FirstWallBuilder.COMPONENT_WALL)
+        wall_xz = xz.get_component(InVesselComponentBuilder.COMPONENT_WALL)
         assert wall_xz is not None
         assert wall_xz.depth == 2
 
     def test_BuilderError_if_wall_does_not_enclose_x_point(self):
         params = copy.deepcopy(self._params)
         params.update({"r_fw_ib_in": 7, "r_fw_ob_in": 9})
-        builder = FirstWallBuilder(
+        builder = InVesselComponentBuilder(
             params, build_config=self._default_config, equilibrium=self.eq
         )
