+++ tests/bluemira/builders/EUDEMO/ivc/test_wall.py
@@ -26,10 +26,10 @@ import copy
 
 import pytest
 
-from bluemira.builders.EUDEMO.first_wall import WallBuilder
+from bluemira.builders.EUDEMO.ivc import WallBuilder
 
 OPTIMISER_MODULE_REF = "bluemira.geometry.optimisation"
-WALL_MODULE_REF = "bluemira.builders.EUDEMO.first_wall.wall"
+WALL_MODULE_REF = "bluemira.builders.EUDEMO.ivc.wall"
 
 CONFIG = {
     "param_class": f"{WALL_MODULE_REF}::WallPolySpline",
